5v Power supply board for Spark Core
A power supply board for the spark core with a wide input voltage range
of 7v to 22v, the footprint is similar to the Spark slightly longer for
access to input terminal. I supplies a regulated 5v output at up to 2A
for driving the spark and other devices.
A header included near the centre of the board passes the input power to
other boards that require the higher voltage